✨ update: 수정 버튼 모듈화 (#186)
